Stabilised BarcodeReader and implemented cleanup.
The BarcodeReader ist now more stable against short transitions, i.e.
between the floor and a paper that is not completely flat and therefore
throws a mini shadow. The cleanup function unregisters the required services
in the timerservice, so the intToBeep() function does work better since
it will not be interrupted as often. The app.c was refactored to
simplify stuff and provide examples.