Latest preparations to set the race.
The only thing that did not work was finding the line after
wallbouncing. No Idea why, though. Hardware was struggling... maybe
because of the heat that day. It worked fine in tests. In case anyone
reads this: Search for another asuro library. This one is more or less
official, but completely retardly designed.